eldercare/chatbot/requirements.txt
python-dotenv>=0.19.0
aiohttp>=3.8.0
google-generativeai>=0.3.0
langchain>=0.0.200
faiss-cpu==1.7.4
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2
tenacity>=8.0.0
backoff>=2.2.0
numpy==1.24.3
azure-ai-inference>=1.0.0
